Fix minimum platform version check
This code lead to "ios 11.4" on my iPad. The platformVersion is used
with QVersionNumber to compare to the minimum platform version of new
app versions, and QVersionNumber fails to parse "ios 11.4".
The platform type is also available already as "platformId", so I think
the version should just be omitted here.

Fixes #12